finalize BBC Kafka producer and consumer integration with MongoDB
