--- news-pipeline/kafka_app/sub_producers/.dockerignore
@@ -1,10 +1,0 @@
-__pycache__/
-*.pyc
-*.pyo
-.git/
-*.env
-*.DS_Store
-*.log
-README.md
-tests/
-# Ignore all files in the root directory except for the Dockerfile
