--- news-pipeline/kafka_app/sub_producers/.gitignore
@@ -1,34 +1,0 @@
-# Ignore all JSON files
-*.json
-
-# Ignore Python bytecode and cache
-__pycache__/
-*.py[cod]
-
-# Ignore logs and scraped URLs
-*.log
-scraped_urls.txt
-bbc_all_articles.json
-
-# macOS system files
-.DS_Store
-
-# VSCode settings
-.vscode/
-
-# Ignore env files if any
-.env
-
-
-# Ignore environment and Docker setup files
-.env
-Dockerfile
-docker-compose.yml
-requirements.txt
-
-# Python junk
-__pycache__/
-*.pyc
-
-# macOS system file
-.DS_Store
