--- news-pipeline/kafka_app/sub_producers/Dockerfile
@@ -1,9 +1,0 @@
-FROM python:3.10-slim
-WORKDIR /app
-
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-CMD ["python", "bbc_producer.py"]
