--- news-pipeline/kafka_app/sub_producers/docker-compose.yml
@@ -1,45 +1,0 @@
-version: "3.8"
-
-services:
-  zookeeper:
-    image: confluentinc/cp-zookeeper:7.2.1
-    container_name: kafka_zookeeper
-    environment:
-      ZOOKEEPER_CLIENT_PORT: 2181
-
-  kafka:
-    image: confluentinc/cp-kafka:7.2.1
-    container_name: kafka_broker
-    depends_on:
-      - zookeeper
-    ports:
-      - "9093:9092"
-    environment:
-      KAFKA_BROKER_ID: 1
-      KAFKA_ZOOKEEPER_CONNECT: kafka_zookeeper:2181
-      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka:9092
-      KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
-
-  mongo:
-    image: mongo
-    container_name: local_mongo
-    ports:
-      - "27018:27017"
-
-  bbc-producer:
-    build: .
-    container_name: bbc_producer
-    depends_on:
-      - kafka
-    command: ["sh", "-c", "sleep 20 && python bbc_producer.py"]
-    env_file:
-      - .env
-
-  bbc-consumer:
-    build: .
-    container_name: bbc_consumer
-    depends_on:
-      - kafka
-    command: ["sh", "-c", "sleep 25 && python bbc_consumer.py"]
-    env_file:
-      - .env
