--- news-pipeline/kafka_app/sub_producers/requirements.txt
@@ -1,5 +1,0 @@
-requests
-beautifulsoup4
-kafka-python
-pymongo
-python-dotenv
